pyproject.toml
[project]
name = "shard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
